Fix a few compiler warnings
git-svn-id: https://svn.osgeo.org/gdal/trunk@25795 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gdal/frmts/iso8211/ddfmodule.cpp
@@ -275,7 +275,7 @@ int DDFModule::Open( const char * pszFilename, int bFailQuietly )
     pachRecord = (char *) CPLMalloc(_recLength);
     memcpy( pachRecord, achLeader, nLeaderSize );
 
-    if( VSIFReadL( pachRecord+nLeaderSize, 1, _recLength-nLeaderSize, fpDDF )
+    if( (int)VSIFReadL( pachRecord+nLeaderSize, 1, _recLength-nLeaderSize, fpDDF )
         != _recLength - nLeaderSize )
     {
         if( !bFailQuietly )
